Refactor base wallaroo.py library into modules
