+++ demos/python_word_count/environment
@@ -1,1 +1,1 @@
-export PYTHONPATH=../../machida/:.
+export PYTHONPATH=../../machida/lib:.
