+++ demos/python_word_count/worker4
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-export PYTHONPATH="../../machida:."
+export PYTHONPATH="../../machida/lib:."
 
 ../../machida/build/machida --application-module word_count --in 127.0.0.1:7010 \
   --out 127.0.0.1:7002 --metrics 127.0.0.1:5001 \
