+++ rules.mk
@@ -44,7 +44,7 @@ wallaroo_path := $(abs_wallaroo_dir)
 integration_path := $(wallaroo_path)/testing/tools
 integration_bin_path := $(integration_path)/integration
 wallaroo_lib :=  $(wallaroo_path)/lib
-wallaroo_python_path := $(wallaroo_path)/machida
+wallaroo_python_path := $(wallaroo_path)/machida/lib
 machida_bin_path := $(wallaroo_path)/machida/build
 
 EMPTY :=
